del .DS_Store and fix libft

+++ srcs/libft/srcs/ft_itoa.c
@@ -6,13 +6,13 @@
 /*   By: kinamura <kinamura@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/18 23:40:26 by kinamura          #+#    #+#             */
-/*   Updated: 2024/06/18 16:38:21 by kinamura         ###   ########.fr       */
+/*   Updated: 2024/07/02 02:51:57 by kinamura         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static size_t	ft_size(long long n)
+static size_t	ft_size(long n)
 {
 	size_t	size;
